requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
python-multipart>=0.0.9
prometheus-client>=0.20.0
httpx>=0.28.1
pyjwt[crypto]>=2.8.0
redis>=5.0.0
